fiz for templates
