+++ myshows/myshowsapp/urls.py
@@ -4,7 +4,11 @@ from django.urls import path, re_path, include
 from rest_framework.routers import DefaultRouter
 from .views.later_watch_shows import LaterWatchShowViewSet
 from .views.full_watched_shows import FullWatchedShowViewSet
-from .views.client_views import *
+from .views.client_views import IndexView, SearchView, DetailView, StartView
+from .views.auth_views import (
+    LoginView, LogoutView, RegisterView, PasswordResetView, 
+    PasswordResetDoneView, PasswordResetConfirmView, PasswordResetCompleteView
+)
 
 
 router = DefaultRouter()
